[monarch] monarch.actor submodule
This is the proposed `monarch.actor` submodule.

The public API is just a flat namespace, determined by `monarch/actor/__init__.py`.

*Everything else* is a private detail, signified by lots of underscores in filenames. That's a little unforunate, but I think draws some good boundaries, and keeps the API simple to the user.

Some principles that should guide us moving forward:
- **Every symbol has at most two locations**: a private location (which should be exactly the same as its location in the source tree), and optionally a public location in the user-facing API.
- We discourage end-users from depending on private locations via underscores, per python convention.

Differential Revision: [D77709116](https://our.internmc.facebook.com/intern/diff/D77709116/)

[ghstack-poisoned]
