+++ python/monarch/_rust_bindings/__init__.pyi
@@ -1,0 +1,8 @@
+# Copyright (c) Meta Platforms, Inc. and affiliates.
+# All rights reserved.
+#
+# This source code is licensed under the BSD-style license found in the
+# LICENSE file in the root directory of this source tree.
+
+# True iff the rust extension was built with the tensor engine feature.
+def has_tensor_engine() -> bool: ...
