+++ python/monarch/_rust_bindings/monarch_hyperactor/alloc.pyi
@@ -73,7 +73,7 @@ class RemoteAllocatorBase:
     def __new__(
         cls,
         world_id: str,
-        initializer: "monarch.allocator.RemoteAllocInitializer",  # pyre-ignore[11] defined in monarch/python/monarch/allocator.py
+        initializer: "monarch.actor._allocator.RemoteAllocInitializer",  # pyre-ignore[11]
         heartbeat_interval: timedelta = timedelta(seconds=5),
     ) -> Self:
         """
