+++ python/monarch/_testing.py
@@ -13,10 +13,10 @@ from contextlib import contextmanager, ExitStack
 from typing import Any, Callable, Dict, Generator, Literal, Optional
 
 import monarch_supervisor
+from monarch.actor._shape import NDSlice
 from monarch.common.client import Client
 from monarch.common.device_mesh import DeviceMesh
 from monarch.common.invocation import DeviceException, RemoteException
-from monarch.common.shape import NDSlice
 from monarch.controller.backend import ProcessBackend
 from monarch.mesh_controller import spawn_tensor_engine
 from monarch.proc_mesh import proc_mesh, ProcMesh
