+++ python/monarch/actor/_shape.py
