+++ python/monarch/bootstrap_main.py
@@ -4,56 +4,17 @@
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
 
-"""
-This is the main function for the boostrapping a new process using a ProcessAllocator.
-"""
+import warnings
 
-import asyncio
-import importlib.resources
-import logging
-import os
-import sys
+warnings.warn(
+    "monarch.bootstrap_main is deprecated, please use from monarch.actor._bootstrap_main instead.",
+    DeprecationWarning,
+    stacklevel=2,
+)
 
-# Import torch to avoid import-time races if a spawned actor tries to import torch.
-import torch  # noqa[F401]
-
-
-async def main():
-    from monarch._rust_bindings.monarch_hyperactor.bootstrap import bootstrap_main
-
-    await bootstrap_main()
-
-
-def invoke_main():
-    # if this is invoked with the stdout piped somewhere, then print
-    # changes its buffering behavior. So we default to the standard
-    # behavior of std out as if it were a terminal.
-    sys.stdout.reconfigure(line_buffering=True)
-    global bootstrap_main
-
-    # TODO: figure out what from worker_main.py we should reproduce here.
-    from monarch.telemetry import TracingForwarder
-
-    if os.environ.get("MONARCH_ERROR_DURING_BOOTSTRAP_FOR_TESTING") == "1":
-        raise RuntimeError("Error during bootstrap for testing")
-
-    # forward logs to rust tracing. Defaults to on.
-    if os.environ.get("MONARCH_PYTHON_LOG_TRACING", "1") == "1":
-        logging.root.addHandler(TracingForwarder(level=logging.DEBUG))
-
-    try:
-        with (
-            importlib.resources.path("monarch", "py-spy") as pyspy,
-        ):
-            if pyspy.exists():
-                os.environ["PYSPY_BIN"] = str(pyspy)
-            # fallback to using local py-spy
-    except Exception as e:
-        logging.warning(f"Failed to set up py-spy: {e}")
-
-    # Start an event loop for PythonActors to use.
-    asyncio.run(main())
+from monarch.actor._bootstrap_main import *  # noqa
 
 
 if __name__ == "__main__":
+    # noqa
     invoke_main()  # pragma: no cover
