+++ python/monarch/common/client.py
@@ -37,6 +37,7 @@ from monarch._rust_bindings.monarch_extension.client import (  # @manual=//monar
     LogLevel,
     WorldState,
 )
+from monarch.actor._shape import NDSlice
 from monarch.common import messages
 from monarch.common.borrows import Borrow, StorageAliases
 from monarch.common.controller_api import LogMessage, MessageResult, TController
@@ -47,7 +48,6 @@ from monarch.common.invocation import DeviceException, RemoteException, Seq
 from monarch.common.recording import flatten_messages, Recording
 
 from monarch.common.reference import Ref, Referenceable
-from monarch.common.shape import NDSlice
 from monarch.common.stream import StreamRef
 from monarch.common.tensor import Tensor
 from monarch.common.tree import tree_map
