+++ python/monarch/common/recording.py
@@ -10,9 +10,9 @@ import traceback
 from collections import defaultdict
 from typing import cast, Dict, Generator, List, NamedTuple, Tuple, TYPE_CHECKING, Union
 
-from monarch.common.reference import Ref
+from monarch.actor._shape import iter_ranks
 
-from monarch.common.shape import iter_ranks
+from monarch.common.reference import Ref
 
 from monarch.common.tensor import InputChecker
 
@@ -21,8 +21,9 @@ from . import messages
 if TYPE_CHECKING:
     from monarch.common.client import Client
 
+from monarch.actor._shape import NDSlice
+
 from .reference import Referenceable
-from .shape import NDSlice
 from .tensor import Tensor
 
 logger = logging.getLogger(__name__)
