+++ python/monarch/controller/backend.py
@@ -13,9 +13,9 @@ import socket
 from abc import ABC, abstractmethod
 from typing import List, NamedTuple, Optional, Sequence, Tuple
 
-from monarch.common import messages
+from monarch.actor._shape import iter_ranks, Slices as Ranks
 
-from monarch.common.shape import iter_ranks, Slices as Ranks
+from monarch.common import messages
 from monarch_supervisor import (
     Context,
     FunctionCall,
