+++ python/monarch/controller/rust_backend/controller.py
@@ -29,11 +29,12 @@ from monarch._rust_bindings.monarch_hyperactor.proc import (  # @manual=//monarc
 )
 
 from monarch._rust_bindings.monarch_messages.debugger import DebuggerAction
+
+from monarch.actor._shape import NDSlice
 from monarch.common.controller_api import LogMessage, MessageResult
 from monarch.common.device_mesh import no_mesh
 from monarch.common.invocation import DeviceException, RemoteException
 from monarch.common.messages import SupportsToRustMessage
-from monarch.common.shape import NDSlice
 from monarch.common.tensor import Tensor
 from monarch.controller.debugger import read as debugger_read, write as debugger_write
 from pyre_extensions import none_throws
