+++ python/monarch/simulator/command_history.py
@@ -12,9 +12,9 @@ from dataclasses import dataclass
 from typing import List, NamedTuple, Optional, Sequence
 
 import torch
+from monarch.actor._shape import NDSlice
 
 from monarch.common import messages
-from monarch.common.shape import NDSlice
 from monarch.simulator.ir import IRGraph
 from monarch.simulator.tensor import DTensorRef
 from monarch.simulator.utils import clean_name, file_path_with_iter
