+++ python/monarch/simulator/interface.py
@@ -6,9 +6,10 @@
 
 from typing import Union
 
+from monarch.actor._shape import NDSlice
+
 from monarch.common.client import Client as _Client
 from monarch.common.device_mesh import DeviceMesh
-from monarch.common.shape import NDSlice
 
 from monarch.simulator.ir import IRGraph
 from monarch.simulator.simulator import (
