+++ python/monarch/world_mesh.py
@@ -8,10 +8,11 @@
 
 from typing import List
 
+from monarch.actor._shape import NDSlice
+
 from monarch.common.client import Client
 
 from monarch.common.device_mesh import DeviceMesh
-from monarch.common.shape import NDSlice
 
 from monarch.controller.backend import ProcessBackend
